# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shopping Cart",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost/api"
        }
    ],
    "paths": {
...
        "/api/promotions": {
            "put": {
                "tags": [
                    "Promotion"
                ],
                "summary": "Update a new promotion",
                "operationId": "91d9b6a5414516de05a6557fba5c3d6a",
                "requestBody": {
                    "description": "Promotion data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "product_code": {
                                        "type": "string",
                                        "example": "ABC123"
                                    },
                                    "rules": {
                                        "type": "string",
                                        "example": "promotion_type:buy_1_get_1 rules:min_quantity:1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Promotion updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Promotion updated successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Failed to create Promotion"
                    }
                }
            },
            "post": {
                "tags": [
                    "Promotion"
                ],
                "summary": "Create a new promotion",
                "operationId": "581f96a26fd7d2b40f30c156dd48699e",
                "requestBody": {
                    "description": "Promotion data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_code": {
                                        "type": "string",
                                        "example": "ABC123"
                                    },
                                    "rules": {
                                        "type": "string",
                                        "example": "promotion_type:buy_1_get_1 rules:min_quantity:1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Promotion created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Promotion created successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Failed to create Promotion"
                    }
                }
            }
        },
        "/api/promotions/{id}": {
            "delete": {
                "tags": [
                    "Promotion"
                ],
                "summary": "Delete a promotion",
                "operationId": "7760da6086d26d538f8c71ce975b9f83",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the promotion to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Promotion deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Deleted success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to delete Promotion"
                    }
                }
            }
        }
    }
}